+++ ui/packages/shared/components/src/index.tsx
@@ -21,6 +21,7 @@ import GrpcMetadataContext, {GrpcMetadataProvider, useGrpcMetadata} from './Grpc
 import Input from './Input';
 import {KeyDownProvider, useKeyDown} from './KeyDownContext';
 import Modal from './Modal';
+import {NoDataPrompt} from './NoDataPrompt';
 import ParcaContext, {defaultValue, ParcaContextProvider, useParcaContext} from './ParcaContext';
 import Pill, {PillVariant} from './Pill';
 import ResponsiveSvg from './ResponsiveSvg';
@@ -31,6 +32,7 @@ import Tab from './Tab';
 import Table from './Table';
 import EmptyState from './EmptyState';
 import {useURLState} from './hooks/useURLState';
+import TextWithTooltip from './TextWithTooltip';
 
 export type {PillVariant, SelectElement};
 
@@ -48,6 +50,7 @@ export {
   Input,
   KeyDownProvider,
   Modal,
+  NoDataPrompt,
   ParcaContext,
   defaultValue as parcaContextDefaultValue,
   ParcaContextProvider,
@@ -57,6 +60,7 @@ export {
   Spinner,
   Tab,
   Table,
+  TextWithTooltip,
   EmptyState,
   useGrpcMetadata,
   useKeyDown,