+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.198](https://github.com/parca-dev/parca/compare/@parca/web@0.16.197...@parca/web@0.16.198) (2023-02-13)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.197](https://github.com/parca-dev/parca/compare/@parca/web@0.16.196...@parca/web@0.16.197) (2023-02-13)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.196](https://github.com/parca-dev/parca/compare/@parca/web@0.16.195...@parca/web@0.16.196) (2023-02-13)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.195](https://github.com/parca-dev/parca/compare/@parca/web@0.16.194...@parca/web@0.16.195) (2023-02-13)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.194](https://github.com/parca-dev/parca/compare/@parca/web@0.16.193...@parca/web@0.16.194) (2023-02-09)
 
 **Note:** Version bump only for package @parca/web